Changed the CSTMAX or maximum number of timesteps to follow a particle to real*8 for cases where it could get larger than would accurately fit in a real Change in mod_interface.f to use a dynamically assigned unit number instead of using 99 which could conflict with other places in the code. Some fixing of deprecated features like real DO loop limits and shared DO loop termination (Goal is to reduce or eliminate compiler warnings).

+++ comsrc/f90/hc_wbc_comp.f90
@@ -96,7 +96,7 @@ Contains
     Integer, Intent (In) :: Num_Pri_Launch ! NPROD
     Integer, Intent (In) :: Num_Sec_Launch ! LPROD
     Integer, Intent (In) :: Number_Produced ! IPROD
-    Real, Intent (In) :: Step_Count ! CIST
+    Real*8, Intent (In) :: Step_Count ! CIST
  
     Check_WBC_Neut_Position = 0
